feat: initial extraction of fold.rs
